Add files via upload
Tried to add the nickname and button for restart the nickname i cant still figure out how to make it work as it needs the cookies so i commented it out for now. the button is there also but doesnt show for some reason also i added some comments and changed the speeds a bit so its a bit easier to move left and right as well. Also found out that this game doesnt keep all the previous player data like the one before since this one doesnt call the new player function from the server so at least thats fixed.